% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meal.R
\name{meal_by_category}
\alias{meal_by_category}
\title{Filter meals by category}
\usage{
meal_by_category(category)
}
\arguments{
\item{category}{Category name (e.g. "Seafood", "Chicken", "Dessert")}
}
\value{
tibble: id, name, thumbnail
}
\description{
Filter meals by category
}
