% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meal.R
\name{meal_context}
\alias{meal_context}
\title{Show TheMealDB package context for LLM use}
\usage{
meal_context()
}
\value{
Invisible string with full context
}
\description{
Show TheMealDB package context for LLM use
}
