% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meal.R
\name{meal_random}
\alias{meal_random}
\title{Get a random meal}
\usage{
meal_random()
}
\value{
tibble with one row: id, name, category, area, instructions, thumbnail, tags, youtube
}
\description{
Get a random meal
}
