% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meal.R
\name{meal_search}
\alias{meal_search}
\title{Search meals by name}
\usage{
meal_search(name)
}
\arguments{
\item{name}{Meal name or partial name (e.g. "chicken", "pasta")}
}
\value{
tibble: id, name, category, area, instructions, thumbnail, tags, youtube
}
\description{
Search meals by name
}
